Update the closed loop interpretation.
Now when a closed loop file is created, the file owner is treated
as the issuer of all configured closed loops. This should make sense
because the omnibus account represents the total value of the
closed loop funds in circulation.

Also added a checkbox to reinterpret the file upon save.